feat: implement iterative learning algorithm for inverse Hammerstein predistortion and remove obsolete challenge scripts
